Changed the '%%%' strings that appear in the files that build the Form.pm files from templates to '!!!'.  '%%' is the comment character for PDF, and while using '%%%' wasn't actually a bug, the potential is too great for comfort.
Added 'local($_)' to all subroutines that use $_.  This fixes a nasty
initialization bug that caused SingletonMap to go off the rails.

Fixed a bug in Bivio::UI::PDF::Form::Form::internal_read_data that caused
it to read off the end of the __DATA__ section and go into an infinite
loop.

Incorporated Rob's code cleanups in Bivio::UI::PDF::Form::f1065/y1999/Form.pm,
which is generated from a template, into the template.